refactor(users): move touch helper to adapter package
